Create README.md
The project is a web application consisting of an online store for selling makeup products, jewelry, and clothing. It is divided into CSS, HTML, and JavaScript components. Each of these languages serves a specific purpose:

HTML (Hypertext Markup Language): It provides the structure for web pages, defining the elements that make up the content.

CSS (Cascading Style Sheets): It determines the presentation and style of the HTML elements, including layout, colors, fonts, and more.

JavaScript: It adds interactivity to the website, allowing for dynamic behavior such as user interactions, animations, and data manipulation.